fix(apps): update models to match API spec
BREAKING CHANGE: Most of the request/response models related to apps will need to be updated.

Anything using `App` will need to be changed to either `AppResponse` (if used with GET)
or `PagedApp` (if used with LIST). AppEndpoints that previously returned `Count` or `Status` now
return an empty Promise.

fix #89